+++ src/crud_views_object_detail/lib/views.py
@@ -1,9 +1,9 @@
 from crud_views.lib.view import CrudViewPermissionRequiredMixin
-from crud_views.lib.views import DetailCustomView  # Task 7 switches this to the renamed DetailView
+from crud_views.lib.views import DetailView
 from crud_views_object_detail.lib.mixins import ObjectDetailMixin
 
 
-class ObjectDetailView(ObjectDetailMixin, DetailCustomView):
+class ObjectDetailView(ObjectDetailMixin, DetailView):
     pass