+++ tests/test1/app/resources.py
@@ -7,7 +7,7 @@ from crud_views.lib.resource import Resource, ResourceViewMixin
 from crud_views.lib.table import Table
 from crud_views.lib.views import (
     ActionViewPermissionRequired,
-    DetailCustomViewPermissionRequired,
+    DetailViewPermissionRequired,
     ListViewPermissionRequired,
     ListViewTableMixin,
     MessageMixin,
@@ -81,7 +81,7 @@ class S3FileListView(ResourceViewMixin, ListViewTableMixin, ListViewPermissionRe
     cv_list_actions = ["detail", "delete", "touch"]
 
 
-class S3FileDetailView(ResourceViewMixin, DetailCustomViewPermissionRequired):
+class S3FileDetailView(ResourceViewMixin, DetailViewPermissionRequired):
     cv_viewset = cv_s3file
     template_name = "app/s3file_detail.html"
 
@@ -178,6 +178,6 @@ class PublisherFileListView(ResourceViewMixin, ListViewTableMixin, ListViewPermi
     cv_list_actions = ["detail"]
 
 
-class PublisherFileDetailView(ResourceViewMixin, DetailCustomViewPermissionRequired):
+class PublisherFileDetailView(ResourceViewMixin, DetailViewPermissionRequired):
     cv_viewset = cv_publisher_file
     template_name = "app/s3file_detail.html"